28/02/2023
Last Push of the Day

+++ app/Http/Controllers/MoveController.php
@@ -179,7 +179,7 @@ class MoveController extends Controller {
                         $this->assignCards($player_cards,$move->user(),$cards_played->cards_played,'remove')),$move);
                     $Game->winner = $this->findWinner($player_cards);
                     $Game->save();
-                    $Room->delete();
+                    !is_null($Room) && $Room->delete();
                 }
                 else {
                     $State =  new GameStateResource($this->newState($move->game(),$Last_State->sequence(),false,
@@ -206,7 +206,7 @@ class MoveController extends Controller {
                         $State = new GameStateResource($this->newState($move->game(),$Last_State->sequence(),true,
                             false,$move->cards(),$this->nextTurn($GamePlayers,$move->user()),'2',['cards_down'=>[]],
                             $this->assignCards($player_cards,$move->user(),$cards_down->cards_down,'add')),$move);
-                        $Room->delete();
+                        !is_null($Room) && $Room->delete();
                     }
                     else {
                         $State = new GameStateResource($this->newState($move->game(),$Last_State->sequence(),true,
@@ -224,13 +224,12 @@ class MoveController extends Controller {
                             $move->user()),'2',['cards_down'=>[]],$player_cards),$move);
                     $Game->winner = $this->findWinner($player_cards);
                     $Game->save();
-                    $Room->delete();
+                    !is_null($Room) && $Room->delete();
                 }
                 else {
                     if($Last_Move->status() === 3) {
                         $State =  new GameStateResource($this->newState($move->game(),$Last_State->sequence(),false,
-                            false,$move->cards(),$this->nextTurn($GamePlayers,
-                                $move->user()),'1',['cards_down'=>[]],$player_cards),$move);
+                            false,$move->cards(), $move->user(),'1',['cards_down'=>[]],$player_cards),$move);
                     }
                     else {
                         $State =  new GameStateResource($this->newState($move->game(),$Last_State->sequence(),false,
